# Use shift operations whenever possible to speedup