This is related to #137 and #139 and replaces PR #140. Using shift operations (in `shift_r8()`) and the (unchanged) `copy_n()` function, but now called with aligned arrays, we speedup all possible operations.  The speedup is visible on large arrays where previously unaligned buffers had to be copied.  This includes insertion, deletion and slicing.  Here are some timings, here `a` is a bitarray of size 2^30 and `b` a bitarray or size 2^27-1:
```
 0.044 sec   insert(0, 1)  little-endian
 0.193 sec   insert(0, 1)  big-endian
 0.087 sec   pop(0)        little-endian
 0.231 sec   pop(0)        big-endian
 0.123 sec   a[1:]         little-endian
 0.209 sec   a[1:]         big-endian
 0.108 sec   b.extend(a)   little-endian
 0.254 sec   b.extend(a)   big-endian
 0.053 sec   b *= 7        little-endian
 0.199 sec   b *= 7        big-endian
```
Notice that little-endian operations are always slower.  The reason for this is that uint64 shift operations need to operate on byte reversed buffers if the buffer represents a big-endian bitarray.  For little-endian bitarrays the byte reversal is not necessary.

Without this PR:
```
 7.695 sec   insert(0, 1)  little-endian
 7.676 sec   insert(0, 1)  big-endian
 6.810 sec   pop(0)        little-endian
 6.800 sec   pop(0)        big-endian
 8.111 sec   a[1:]         little-endian
 8.178 sec   a[1:]         big-endian
 9.236 sec   b.extend(a)   little-endian
 9.269 sec   b.extend(a)   big-endian
 6.816 sec   b *= 7        little-endian
 6.926 sec   b *= 7        big-endian
```